+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,38 @@
+import path from 'path';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+import moduleA from './files/a.json' assert { type: 'json' };
+import moduleB from './files/b.json' assert { type: 'json' };
+import cModule from './files/c.js';
+
+cModule()
+
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = moduleA
+} else {
+    unknownObject = moduleB
+}
+console.log(`random: ${random}`)
+console.log(unknownObject)
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const createMyServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+export { unknownObject, createMyServer}
