+++ src/modules/files/c.js
@@ -1,1 +1,3 @@
-console.log('Hello from c.js!');
+export default function hello() {
+  return console.log('Hello from c.js!');
+}
